Removed .sty extension / absolute paths for cite
Removed the .sty extension, which caused some confusion for the root
detection routine. Not entirely sure what caused the issue, but gf to
.sty seems like a border-line enough case anyway.

Also, I've modified the bibsearch function to better work with larger
projects. It prepends the path of whichever file links to a bib file
to that bib file's path, allowing us to correctly complete citations
even if the bib file and current file are in two different directories
and the main file linking to the bibfile is in a different directory
to the current file.

This could possibly be modified to check for absolute paths in
\bibliography commands, but I don't believe anyone is using those?