Improved latexmk interface: added callback
Added a simple callback function that utilizes the clientserver in vim and the
$success_cmd and $failure_cmd in latexmk to call latex#latexmk#error() after
compilation.  The feature is enabled by default, because it strikes me as
a very nice feature!

Note that this feature might require a minor update to existing .latexmkrc
files, in that existing $success_cmd and $failure_cmd should end in a semicolon
in order to allow vim-latex to append the callback.  I hope I might remove this
restriction in a later update, but I didn't see how to do this now.