Edited fix used for bug #191
The previous method used to determine if the foldmethod option was set from
a modeline wasn't working for me.

The function vimtex#fold#s:check_modeline was always returning false.

I changed the function to use a simple regex search on the actual file
itself searching for a modeline in the proper line ranges.