Fixed #440: Problem with stopline arguments
The stoplines were relative to the window limits (i.e. line('w$') and
line('w0')). This does not work when delimiters are searched beyond the
window limit +- the stopline borders, which e.g. might happen when
searching in large documents.

The fix is to use limits that are relative to the current position,
line('.').