Added notice for terminal users in readme's installation section
Hello, I know this is mentioned in the documentation of this program, but it gave me quite a hustle to figure out for a first time since I wasn't using any plugins that required `+clientserver` functionality of vim until now, so I thought that vimtex could be at cause here rather than some vim ambiguity. So to clarify things I think this deserves spot in readme since it's actually an installation step for some use cases.