Changed the behavior of create command function
The `vimtex#cmd#create()` has been modified according to [this discussion](https://github.com/lervag/vimtex/pull/596) prompted by a previous pull request concerning edits to the same function.

Note. I don't know whether it is possible or not to make a regex stop at the current position of the cursor, so I haven't modified that line with the regex, but the two lines where `l:strpart2` and `l:strpart3` are defined (in particular I've extracted a part of `l:match`).

I've not touched the position of the cursor after the LaTeX command is inserted, since I think it's the best position already.