Delete begin/end of environment along with lines if they remain empty
A change is made to vimtex#env#delete in a way that it deletes `\begin{...}` and `\end{...}` along with the lines they belong to if they're the only content of those lines. Before this change the command dse could leave empty lines in place of `begin`/`end` lines. The commit is coherent with the discussion [here](https://github.com/lervag/vimtex/pull/703).