use 'menu' field only for candidate-specific info
For type of candidate (cmd, env, etc.), use 'kind' field.

This allows ncm2 not to match on 'kind' field to avoid polluting
the candidate list. (The only possible drawback is that the list
of commands cannot be filtered by the package they come from; if
this is desired, one can add a separate matcher for 'kind'.)

As a bonus (and the only visible change when *not* using ncm2),
the unicode symbols of command candidates is now aligned across
packages, which (in my opinion) looks nicer).