doc: register ncm2 source on syntax set
If nvim is started with a non-tex buffer, ncm2 is loaded but the
vimtex regex is not available, giving an error. The updated config
only registers the sources when `Syntax` is set to `tex` (which is
the scope of the source anyway), in which case vimtex should be
loaded already.