Parse $pdf_mode variable in latexmkrc
The $pdf_mode variable determines the compiler used by latexmk when
compiling a pdf file. Its possible values, described in the latexmk man
page, are:

    0   don't generate a pdf file       [ignored]
    1   use pdflatex
    2   generate ps, then use ps2pdf    [ignored]
    3   generate dvi, then use dvipdf   [ignored]
    4   use lualatex
    5   use xelatex

The options marked as [ignored] are not taken into account by the new
init_pdf_mode_option() function.

If the value of $pdf_mode is 1, 4 or 5, the value of tex_program is
overriden by the corresponding value in the latexmkrc file, in a
similar way as it was done for the build_dir option.