Autoclose the QuickFix window (default: off)
The integer option `g:vimtex_quickfix_autoclose_after_keystrokes` sets
the initial number for a countdown of cumulative movements, edits, etc
(counted in terms of occurrences of `CursorMoved,CursorMovedI` events)
at the end of which the QuickFix window automatically closes if the user
has not jumped into it. The countdown is stopped and reset: if the user
jumps back out of the QuickFix window, the countdown starts all over
again. The described behavior does not depend on
`g:vimtex_quickfix_mode`. The new option is off by default
(`g:vimtex_quickfix_autoclose_after_keystrokes == 0`).

The countdown is delegated to the new function `s:qf_autoclose_check`,
which is called by an `autocmd` defined in an `augroup`, the latter
created whenever the QuickFix window is opened and deleted by the
aforementioned function itself when the countdown reaches zero (_i.e._
when the QuickFix window autocloses).

An `autocmd!` in the `augroup` ensures no "lingering" `autocmd`s.

A script-local variable is used as counter.