VimtexDocPackage: add support for environments
Previously calling VimtexDocPackage with the cursor over `\begin{foo}`
showed documentation related to \begin. Showing documentation
related to foo is probably always more useful.

As the environments are stored in the same completion file as the
commands (in the format ^\begin{foo}$), we can mostly wrap the code for
commands to get nice documentation for environments.