Remove texZone link for Asymptote syntax
Because texAsymptoteZone includes the asy syntax file, linking it to
texZone makes otherwise unassigned parts of the code (e.g. variable
names) to be in the texZone color instead of the default color; this
commit removes this link so that Asymptote code is highlighted as if it
were in a standalone .asy file.