Do not parse .blg files if bibunits is used.
The bibunits [1] package provides a way to use multiple bibliographies
in one document.

A .bib file can be included using bibunits via the commands \putbib[...]
or \defaultbibliography{...}.  Vimtex can be told about these .bib files
by changing the g:vimtex_bibliography_commands variable.

However, bibunits generates separate .blg files (sequentially named
bu1.blg, bu2.blg, etc.) for each bibliography and because of this, the
main .blg file does not contain any information about the .bib files
that are actually used.  But it is the main .blg file that Vimtex parses
and since there is nothing in it, Vimtex cannot find the right .bib
files.  This in turn prevents omni completion of the bibliographic
entries.

Hence, if bibunits is used, one should resort to manual parsing of the
document using the s:files_manual() function to find the right .bib
files.

[1]: https://www.ctan.org/pkg/bibunits