Backward search: better support for tab pages
When a file is open in a different tab page, make sure to switch to that
tab page when SyncTeX backward search requests that file. The previous
behavior was to use bufwinnr, which does not look at windows in other
tab pages and therefore acts as if the buffer is not open anywhere.

Refer https://stackoverflow.com/a/42628934