indentkeys autoindent only at start of line
This PR prefixes the 'indentkeys' `(,),],},\item,\else,\fi` with a `0`. 
The effect is that these indent keys only autoindent the line when they are typed at the start of the line (see `:h i_CTRL-F`).
ie. no autoindenting when typing a closing `),],}` mid-way through a statement.

(The assumption here is that one would *not* want an autoindent when, say, is in the midst of a long equation,
but only to autoindent when closing an existing block).