Add tentative section for third-party guides
Consider this the first draft of a section for third-party VimTeX guides
on the larger Web. As we discussed earlier, [@lervag](https://github.com/lervag), I have (1) referenced and briefly described my series and (2) invited the VimTeX community to contribute other (good-quality) third-party content.

As in commit 58e704a, feel free to edit this section as you see fit.