doc: rename "recursive search" to "directory scan"
The method 6 of finding the main file is based on a directory scan which
is described as "recursive" scan. However, while a recursive scan
suggests a (recursive) directory descent, the opposite is true: the
unique ascending path of parent directories is scanned.

This is fixed in the following steps:

  1. The method is renamed to "directory scan method".
  2. The method description is adapted to reflect this behavior more
     clearly. In particular, it is explicitly mentioned that no
     recursive directory descent is performed.
  3. The option to disable this method is renamed to
     g:vimtex_disable_directory_scan_main_file_detection
  4. Some vimscript comments are fixed that actually mention an upwards
     recursion in the directory tree.

As a side effect, the overall description of the directory scan method
is reworked.

As a further side effect, the name of this method is more independent of
the algorithmic mechanism, but rather reflects the policy, cf. "policy
versus mechanism" paradigm