feat!: change build_dir to out_dir
As a step towards supporting aux_dir for latexmk I first want to
consolidate on the out_dir name which is used for several backends. The
term "build_dir" is not really used in these tools, so no need to add
a new term here. Instead, we revert to out_dir, which is a specific
concept in both latexmk, latexrun and tectonic. Further, the current
implementation of these things is really quite similar to the
"build_dir" concept.

refer: #2534