perf(syntax): Rewrite skip regexes
`:syntime report` shows that the "skip" rules for `texMathZoneTI` and
`texMathGroup` are very expensive. The change might speed things up a bit
because every non-backslash character only needs to be looked at once.
According to `:syntime report` this reduces the time spent on the
`texMathZoneTI` skip rule from 3.1% to 2.7%, and similar for texMathGroup.
Also change other skip rules of the form `\\<char 1>\|\\<char 2>`.

Overall, this leads to a ~1% speedup.