perf(syntax): Re-order math syntax rules
Re-order means change the order in which they are checked. Most
importantly, (using the `_texMathBackslash` rule) we try to make the
syntax engine check syntax rules for commands first when we find a
backslash (instead of looking for things like `[(){}+-=]` etc.)

This leads to a ~5% speedup.

This is definitely suboptimal, because it's one more syntax rule, when
we could in theory just move the definitions of
```
texComment,
texSpecialChar,
texMathSymbol,
texCmdGreek,
texTabularChar,
texCmdEnvM,
texCmdFootnote,
texCmdMinipage,
texCmdParbox,
texCmdRef,
texCmdSize,
texCmdStyle,
texCmdTodo,
texCmdVerb,
texMathCmd,
texMathCmdEnv,
texMathCmdStyle,
texMathCmdStyleBold,
texMathCmdStyleItal,
texMathCmdText,
texMathDelimMod,
texMathSymbol,
```
to the bottom of `vimtex#syntax#core#init_rules()` to get the same effect.

Author: Carlos Esparza <ca2010es@gmail.com>