fix: simplify root handling during tex parser
Did not see exactly what was wrong, but it seems I was passing the root
as part of an option dict and somehow did not control the dict properly.
But we did not need the dict and could just simply pass the root value
directly. That seems to fix things.

refer: #3208